django_backend/tests/requirements.txt
# Test dependencies
pytest>=7.4.0
pytest-django>=4.7.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.24.0

